Move tests to offline module. Now they tests flow version of QueryChannelsController
