+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/querychannels/QueryChannelsControllerTests.java
@@ -1,0 +1,5 @@
+package io.getstream.chat.android.offline.querychannels;
+
+class QueryChannelsControllerTests {
+
+}
